Do not return the size when writing the file
The written size is being mistakenly returned in the RadosFsFile
and RadosFsIO's write methods. Instead, just the error or success
code should be returned.

+++ src/RadosFsFile.hh
@@ -57,9 +57,9 @@ public:
 
   ssize_t read(char *buff, off_t offset, size_t blen);
 
-  ssize_t write(const char *buff, off_t offset, size_t blen);
+  int write(const char *buff, off_t offset, size_t blen);
 
-  ssize_t writeSync(const char *buff, off_t offset, size_t blen);
+  int writeSync(const char *buff, off_t offset, size_t blen);
 
   int create(int permissions = -1);
