+++ src/RadosFsIO.hh
@@ -39,8 +39,8 @@ public:
   ~RadosFsIO();
 
   ssize_t read(char *buff, off_t offset, size_t blen);
-  ssize_t write(const char *buff, off_t offset, size_t blen);
-  ssize_t writeSync(const char *buff, off_t offset, size_t blen);
+  int write(const char *buff, off_t offset, size_t blen);
+  int writeSync(const char *buff, off_t offset, size_t blen);
 
   std::string inode(void) const { return mInode; }
